requirements.txt
flask
python-dotenv
langchain
langchain-groq
faiss-cpu
huggingface-hub
transformers
langchain-community
groq
sentence-transformers
